move a bunch of section generator functions to their own file
